refactor out components
